requirements.txt
psycopg2-binary
pandas
click==8.1.8
colorama==0.4.6
joblib==1.4.2
nltk==3.9.1
numpy==2.2.3
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.1
regex==2024.11.6
six==1.17.0
tqdm==4.67.1
tzdata==2025.1
Unidecode==1.3.8
nltk